Changed "double" to "float" to use single precision instead of double precision.

+++ include/dataParsers/dataParser.cuh
@@ -18,7 +18,7 @@ class DataParser {
     protected:
         std::string getFilePath();
         int getType();
-        double parseGroups(std::string groupString);
+        float parseGroups(std::string groupString);
 
     public:
         DataParser(std::string filePath, int type);
