+++ include/positions/callPosition.cuh
@@ -6,10 +6,10 @@
 class CallPosition : public Position {
     public:
         __device__ __host__ CallPosition() : Position() {}
-        __device__ __host__ CallPosition(const char *symbol, int timestamp, double price, double investment, double profitability, int expirationMinutes)
+        __device__ __host__ CallPosition(const char *symbol, int timestamp, float price, float investment, float profitability, int expirationMinutes)
             : Position(symbol, timestamp, price, investment, profitability, expirationMinutes) {}
         __device__ __host__ ~CallPosition() {}
-        __device__ __host__ double getProfitLoss();
+        __device__ __host__ float getProfitLoss();
 };
 
 #endif
