+++ include/studies/rsiStudy.cuh
@@ -10,13 +10,13 @@
 class RsiStudy : public Study {
     private:
         int dataSegmentLength;
-        double previousAverageGain;
-        double previousAverageLoss;
+        float previousAverageGain;
+        float previousAverageLoss;
 
     public:
-        RsiStudy(std::map<std::string, double> inputs, std::map<std::string, std::string> outputMap);
-        double calculateInitialAverageGain(Tick *initialTick, std::vector<Tick*> *dataSegment);
-        double calculateInitialAverageLoss(Tick *initialTick, std::vector<Tick*> *dataSegment);
+        RsiStudy(std::map<std::string, float> inputs, std::map<std::string, std::string> outputMap);
+        float calculateInitialAverageGain(Tick *initialTick, std::vector<Tick*> *dataSegment);
+        float calculateInitialAverageLoss(Tick *initialTick, std::vector<Tick*> *dataSegment);
         void tick();
 };
