+++ include/types/configurationOption.cuh
@@ -1,7 +1,7 @@
 #ifndef CONFIGURATIONOPTION_H
 #define CONFIGURATIONOPTION_H
 
-typedef std::vector<std::map<std::string, boost::variant<std::string, double, int>>> ConfigurationOption;
+typedef std::vector<std::map<std::string, boost::variant<std::string, float, int>>> ConfigurationOption;
 
 #endif
