+++ include/types/mapConfiguration.cuh
@@ -1,7 +1,7 @@
 #ifndef MAPCONFIGURATION_H
 #define MAPCONFIGURATION_H
 
-typedef std::map<std::string, boost::variant<int, double>> MapConfiguration;
+typedef std::map<std::string, boost::variant<int, float>> MapConfiguration;
 
 #endif
