+++ include/types/tick.cuh
@@ -1,6 +1,6 @@
 #ifndef TICK_H
 #define TICK_H
 
-typedef std::map<std::string, double> Tick;
+typedef std::map<std::string, float> Tick;
 
 #endif
