+++ lib/studies/smaStudy.cu
@@ -3,8 +3,8 @@
 void SmaStudy::tick() {
     std::vector<Tick*> *dataSegment = nullptr;
     int dataSegmentLength = 0;
-    double sum = 0.0;
-    double sma = 0.0;
+    float sum = 0.0;
+    float sma = 0.0;
 
     resetTickOutputs();
